feat: major overhaul of memory layout with Buddy System and Main-thread Idle-ization
Summary of Changes:

1. Physical Memory Management: From Bitmaps to Buddy System
   - Implemented a Buddy System (buddy.c) for page-level management, replacing
     the legacy physical memory pool bitmaps. Physical memory is now fully
     managed by the Buddy System, decoupling it from fixed-size bitmaps.
   - Introduced 'global_pages' to store page structs. This array is placed
     immediately after the 1MB reserved kernel space and page tables.
   - Current palloc/pfree Integration: Note that while connected to the
     Buddy System, they currently operate primarily on single-page allocations.
     Full multi-page contiguous allocation benefits are yet to be exploited.
   - Small memory allocation (2B-1024B) still relies on the existing Slab/Arena
     logic, with plans to refactor 'malloc' into a library function using
     'mmap' and 'brk' syscalls.

2. Memory Layout Evolution & Breaking the 128MB Limit:
   - Legacy Constraints: Previously, bitmaps for kernel/user physical pools and
     the kernel virtual pool were crammed into the 0x9a000-0x9dfff region (4 pages).
     This 16KB bitmap capacity limited the theoretical max RAM to ~128MB.
   - New Layout Design:
     - Removed physical pool bitmaps entirely.
     - Relocated the Main thread PCB (formerly at 0x9e000) to the dynamic
       kernel heap area (after_init).
     - Freed up the 0x9a000-0x9efff range (20KB) exclusively for the kernel
       virtual address bitmap, now capable of tracking the full 4GB space.
   - Dynamic Heap Calculation: 'kernel_heap_start' is now a global variable
     calculated as: (1MB Reserved + Page Tables + global_pages size),
     ensuring optimal use of RAM.
   - Result: Confirmed system stability with 3072MB (3GB) RAM via QEMU.
     The 3.5GB+ crash is likely due to x86 architectural holes (MMIO/VGA/BIOS).

3. Initialization Flow & Threading Refactor:
   - Split Initialization: Bifurcated 'init_all' into 'early_init' (pre-heap)
     and 'after_init' (post-heap) to handle the migration of the Main thread.
   - PID Re-alignment: Changed PID start from 1 to 0.
     - Main thread: PID 0.
     - Temporary Idle thread: PID 1.
     - Init process: Now correctly inherits PID 1 after the temp Idle is destroyed.

4. Performance: The "Idle-ization" of Main:
   - Main Thread Transformation: In the final stage of 'after_init', the
     original bootstrap Idle thread is released. The Main thread (PID 0)
     is converted into the permanent Idle thread.
   - Rationale: A temporary Idle thread is kept during early 'after_init'
     to prevent system deadlocks when Main blocks on locks before the
     scheduler is fully ready.
   - Architectural Benefits:
     - Efficiency: PID 0 now utilizes 'hlt' and 'sti' when idle, replacing
       the power-hungry 'while(1)' loop.
     - Speed: Significant throughput boost observed. 'untar_all' execution
       time dropped from ~5s to <1s.
     - Standard Compliance: With PID 0 as Idle and PID 1 as Init, the process
       tree now closely aligns with UNIX standards.

5. Maintenance Note:
   - 'sys_free_mem' has been temporarily restricted to reporting total pool
     capacities. Precise free-page tracking will be re-integrated after
     the Slab system is refactored to interface with the Buddy System.

+++ prog/shell/shell.c
@@ -85,10 +85,10 @@ static void cmd_execute(uint32_t argc, char** argv, int32_t pgid_to_set, bool is
             setpgid(pid, pid);
             if (!is_background) {
                 // 前台：移交 TTY 权限并阻塞等待
-                ioctl(stdin_no, TIOCSPGRP, &pid);
+                ioctl(stdin_no, TIOCSPGRP, (uint32_t)&pid);
                 waitpid(pid, NULL, 0);
                 int32_t shell_pid = getpid();
-                ioctl(stdin_no, TIOCSPGRP, &shell_pid);
+                ioctl(stdin_no, TIOCSPGRP, (uint32_t)&shell_pid);
             } else {
                 // 后台：直接返回，控制权留在 Shell
                 printf("[Background] %d\n", pid);
@@ -149,11 +149,11 @@ int main(void) {
 
             if (!is_background) {
                 // 前台管道，Shell 移交 TTY 给整个进程组并等待
-                ioctl(stdin_no, TIOCSPGRP, &pid1);
+                ioctl(stdin_no, TIOCSPGRP, (uint32_t)&pid1);
                 waitpid(pid1, NULL, 0);
                 waitpid(pid2, NULL, 0);
                 int32_t shell_pid = getpid();
-                ioctl(stdin_no, TIOCSPGRP, &shell_pid);
+                ioctl(stdin_no, TIOCSPGRP, (uint32_t)&shell_pid);
             } else {
                 // 后台管道，打印组长 PID，Shell 继续
                 printf("[Background Pipe Group] %d\n", pid1);
